Extract layouts.py from _compute/runtime.py
Replace the layouts.py re-export shim with the 33 relocated symbol bodies:
the MechanismLayout / ClampRoutingTable records, the layout grouping rules
(choose_layout, mechanism_kind, mechanism_signature, _fn_fingerprint), the
state-buffer allocation helpers, and the clamp / NetStim evaluators.

All 33 blocks are byte-identical to their previous text in runtime.py; this
is a pure relocation with no behavior change. layouts.py is now a leaf: it
imports nothing from runtime.py, and defers its six CellRuntimeState
annotations through `from __future__ import annotations` plus a
TYPE_CHECKING import of .state.

runtime.py drops the imports the move orphaned (warnings, fields,
is_dataclass, Literal, jnp, _to_hashable, and ten braincell.mech names),
gains a back-import block re-exporting the moved names so its import
surface is unchanged, and gets a module docstring describing what remains.
NETSTIM_KIND and CellRuntimeState stay behind for a later task.
